refector: 코드리뷰반영(분리제외)
