--- three session/test/src/styles/theme.jsx
@@ -1,16 +1,0 @@
-export const theme = {
-    colors: {
-      primary: '#FFC1C1',
-      secondary: '#FFF3F3',
-      text: '#333',
-      modalBackground: 'rgba(0, 0, 0, 0.5)',
-      white: 'white',
-      shadow: 'rgba(0,0,0,0.1)'
-    },
-    fonts: {
-      main: 'Comic Sans MS, Arial, sans-serif'
-    }
-  };
-  
-  export default theme;
-
